requirements.txt
Django==5.0.3
djangorestframework==3.14.0
pillow==10.2.0
psycopg2==2.9.9